Fix compilation issues and prepare for GitHub upload
- Update Rust toolchain to latest stable (1.89.0) to resolve edition2024 compatibility
- Fix test compilation errors in call_bridging_tests.rs and rfc_compliance_tests.rs
- Fix type casting errors and struct field mismatches in test files
- Update gitignore to exclude temporary files and build artifacts
- Format code to pass CI formatting checks
- Resolve dependency version conflicts in Cargo.lock
- Add new API modules and web UI components
- Fix various clippy warnings and improve code quality

All core functionality compiles successfully and CI pipeline commands pass.
Some test failures expected in CI environment due to database connectivity.

+++ tests/call_bridging_tests.rs
@@ -279,7 +279,7 @@ impl SipEndpoint {
         };
 
         let mantissa = (sample >> (exponent + 4)) & 0xF;
-        (sign | (exponent << 4) | mantissa) ^ 0x55
+        ((sign | (exponent << 4) | mantissa) ^ 0x55) as u8
     }
 }
